Use conda-forge's git
Seems that we are running into issues cloning a repo into a directory on
Heroku. This seems like a rather basic thing to do with `git`. So it may
just be that Heroku has a really old copy of `git`. This adds `git` from
`conda-forge` so that a recent enough version can be used and this issue
resolved.