Force Python 3.5
As it appears some dependencies are not correctly satisfied on Python
3.6 and they are causing us some issues with getting proper
re-renderings and it appears setting the Miniconda version was not
sufficient to force the Python version, simply pin the Python during
upgrade to 3.5. At least on Python 3.5, we are confident that we have
all the dependencies needed to do proper re-renderings. We can always
revert this pinning once the underlying issues are fixed.